Refactor NamedTemporaryFile call + some cleanup
- Pulled NamedTemporaryFile call out of context manager to handle
  normally

- Added 'extension' optional arg to shell_run_command

- Moved logging fixtures to conftest.py

- Updated "*_error_windows" tests to reflect changes to prefect flow
  definition

- Provided reasons for skipif's across tests
