+++ tests/test_commands_windows.py
@@ -1,7 +1,7 @@
 import logging
 import os
 import sys
-import re
+import glob
 
 import pytest
 from prefect import flow
@@ -9,48 +9,22 @@ from prefect import flow
 from prefect_shell.commands import shell_run_command
 
 
-@pytest.fixture
-def prefect_caplog(caplog):
-    logger = logging.getLogger("prefect")
-
-    # TODO: Determine a better pattern for this and expose for all tests
-    logger.propagate = True
-
-    try:
-        yield caplog
-    finally:
-        logger.propagate = False
-
-
-@pytest.fixture
-def prefect_task_runs_caplog(prefect_caplog):
-    logger = logging.getLogger("prefect.task_runs")
-
-    # TODO: Determine a better pattern for this and expose for all tests
-    logger.propagate = True
-
-    try:
-        yield prefect_caplog
-    finally:
-        logger.propagate = False
-
-
 # TODO: prefect futures result not being returned here, getting AttributeError: 'str' object has no attribute 'result'
-# @pytest.mark.skipif(sys.platform != "win32", reason="")
-# def test_shell_run_command_error_windows(prefect_task_runs_caplog):
-#     @flow
-#     def test_flow():
-#         return shell_run_command(command="ls this/is/invalid", return_all=True, shell="powershell")
+@pytest.mark.skipif(sys.platform != "win32", reason="see test_commands.py")
+def test_shell_run_command_error_windows(prefect_task_runs_caplog):
+    @flow
+    def test_flow():
+        return shell_run_command(command="ls this/is/invalid", return_all=True, shell="powershell")
 
-#     homedir = os.environ["USERPROFILE"]
-#     match = f"ls : Cannot find path {homedir}\\this\\is\\invalid because it does not exist."
-#     with pytest.raises(RuntimeError, match=match):
-#         test_flow().result(raise_on_failure=True)
+    homedir = os.environ["USERPROFILE"]
+    match = f"ls : Cannot find path {homedir}\\this\\is\\invalid because it does not exist."
+    with pytest.raises(RuntimeError, match=match):
+        test_flow()
 
-#     assert len(prefect_task_runs_caplog.records) == 0
+    assert len(prefect_task_runs_caplog.records) == 0
 
 
-@pytest.mark.skipif(sys.platform != "win32", reason="")
+@pytest.mark.skipif(sys.platform != "win32", reason="see test_commands.py")
 def test_shell_run_command_windows(prefect_task_runs_caplog):
     prefect_task_runs_caplog.set_level(logging.INFO)
     echo_msg = "_THIS_ IS WORKING!!!!"
@@ -66,7 +40,7 @@ def test_shell_run_command_windows(prefect_task_runs_caplog):
     assert echo_msg in prefect_task_runs_caplog.text.replace("\r\n\n", "").replace("\r\n", " ")
 
 
-@pytest.mark.skipif(sys.platform != "win32", reason="")
+@pytest.mark.skipif(sys.platform != "win32", reason="see test_commands.py")
 def test_shell_run_command_stream_level_windows(prefect_task_runs_caplog):
     prefect_task_runs_caplog.set_level(logging.WARNING)
     echo_msg = "_THIS_ IS WORKING!!!!"
@@ -87,7 +61,7 @@ def test_shell_run_command_stream_level_windows(prefect_task_runs_caplog):
     assert echo_msg in prefect_task_runs_caplog.text.replace("\r\n\n", "").replace("\r\n", " ")
 
 
-@pytest.mark.skipif(sys.platform != "win32", reason="")
+@pytest.mark.skipif(sys.platform != "win32", reason="see test_commands.py")
 def test_shell_run_command_helper_command_windows():
     @flow
     def test_flow():
@@ -96,7 +70,7 @@ def test_shell_run_command_helper_command_windows():
     assert test_flow() == os.path.expandvars("$USERPROFILE")
 
 
-@pytest.mark.skipif(sys.platform != "win32", reason="")
+@pytest.mark.skipif(sys.platform != "win32", reason="see test_commands.py")
 def test_shell_run_command_return_all():
     @flow
     def test_flow():
@@ -107,7 +81,7 @@ def test_shell_run_command_return_all():
     assert result[1].rstrip() == "yes!"
 
 
-@pytest.mark.skipif(sys.platform != "win32", reason="")
+@pytest.mark.skipif(sys.platform != "win32", reason="see test_commands.py")
 def test_shell_run_command_no_output_windows():
     @flow
     def test_flow():
@@ -116,7 +90,7 @@ def test_shell_run_command_no_output_windows():
     assert test_flow() == ""
 
 
-@pytest.mark.skipif(sys.platform != "win32", reason="")
+@pytest.mark.skipif(sys.platform != "win32", reason="see test_commands.py")
 def test_shell_run_command_uses_current_env_windows():
     @flow
     def test_flow():
@@ -126,7 +100,7 @@ def test_shell_run_command_uses_current_env_windows():
     assert result[0].rstrip() == os.environ["USERPROFILE"]
 
 
-@pytest.mark.skipif(sys.platform != "win32", reason="")
+@pytest.mark.skipif(sys.platform != "win32", reason="see test_commands.py")
 def test_shell_run_command_update_current_env_windows():
     @flow
     def test_flow():
@@ -142,7 +116,7 @@ def test_shell_run_command_update_current_env_windows():
     assert result[0] == os.environ["USERPROFILE"]
     assert result[1] == "test value"
 
-@pytest.mark.skipif(sys.platform != "win32", reason="")
+@pytest.mark.skipif(sys.platform != "win32", reason="see test_commands.py")
 def test_shell_run_command_ensure_suffix_ps1():
     @flow
     def test_flow():
@@ -151,14 +125,13 @@ def test_shell_run_command_ensure_suffix_ps1():
     result = test_flow()
     assert result == "2"
 
-@pytest.mark.skipif(sys.platform != "win32", reason="")
+@pytest.mark.skipif(sys.platform != "win32", reason="see test_commands.py")
 def test_shell_run_command_ensure_tmp_file_removed():
     @flow
     def test_flow():
         temp_dir = os.environ["TEMP"]
-        return shell_run_command(command="echo 'clean up after yourself!'", shell="powershell", dir=f"{temp_dir}\\prefect-cleanup")
+        return shell_run_command(command="echo 'clean up after yourself!'", shell="powershell")
     
     result = test_flow()
     temp_dir = os.environ["TEMP"]
-    test_dir = f"{temp_dir}/prefect-cleanup"
-    assert os.path.exists(test_dir) == False
+    assert len(glob.glob(f"{temp_dir}\\prefect-*.ps1")) == 0
